Ver 2.4 data models
create the data models
