+++ application/routes.py
@@ -1,5 +1,6 @@
 from application import app, db
 from flask import render_template, request, json, Response
+from application.models import User, Course, Enrollment
 
 
 
@@ -49,13 +50,6 @@ def api(idx=None):
     return Response(json.dumps(jdata), mimetype="application/json")
 
 
-#  User Class for Data Base ver 2.01
-class User(db.Document):
-    user_id     =   db.IntField( unique=True )
-    first_name  =   db.StringField( max_length=50 )
-    last_name   =   db.StringField( max_length=50 )
-    email       =   db.StringField( max_length=30 )
-    password    =   db.StringField( max_length=30 )
 
 # User Route to Database
